+++ layers/Schema/packages/comments-wp/src/TypeAPIs/CommentTypeAPI.php
@@ -8,7 +8,7 @@ use PoP\ComponentModel\TypeAPIs\InjectedFilterDataloadingModuleTypeAPITrait;
 use PoP\Hooks\HooksAPIInterface;
 use PoPSchema\Comments\ComponentConfiguration;
 use PoPSchema\Comments\Constants\CommentTypes;
-use PoPSchema\Comments\Constants\Status;
+use PoPSchema\Comments\Constants\CommentStatus;
 use PoPSchema\Comments\TypeAPIs\CommentTypeAPIInterface;
 use PoPSchema\QueriedObject\Helpers\QueriedObjectHelperServiceInterface;
 use PoPSchema\SchemaCommons\DataLoading\ReturnTypes;
@@ -24,10 +24,10 @@ class CommentTypeAPI implements CommentTypeAPIInterface
     public const HOOK_QUERY = __CLASS__ . ':query';
 
     protected array $cmsToPoPCommentStatusConversion = [
-        'approve' => Status::APPROVED,
-        'hold' => Status::ONHOLD,
-        'spam' => Status::SPAM,
-        'trash' => Status::TRASH,
+        'approve' => CommentStatus::APPROVE,
+        'hold' => CommentStatus::HOLD,
+        'spam' => CommentStatus::SPAM,
+        'trash' => CommentStatus::TRASH,
     ];
 
     protected array $popToCMSCommentStatusConversion;
@@ -75,7 +75,14 @@ class CommentTypeAPI implements CommentTypeAPIInterface
 
         // Convert the parameters
         if (isset($query['status'])) {
-            $query['status'] = $this->convertCommentStatusFromPoPToCMS($query['status']);
+            if (is_array($query['status'])) {
+                $query['status'] = array_map(
+                    [$this, 'convertCommentStatusFromPoPToCMS'],
+                    $query['status']
+                );
+            } else {
+                $query['status'] = $this->convertCommentStatusFromPoPToCMS($query['status']);
+            }
         }
         if (isset($query['types'])) {
             $query['type__in'] = $query['types'];
@@ -237,6 +244,19 @@ class CommentTypeAPI implements CommentTypeAPIInterface
         /** @var WP_Comment */
         $comment = $comment;
         return $comment->comment_type;
+    }
+    public function getCommentStatus(object $comment): string
+    {
+        /** @var WP_Comment */
+        $comment = $comment;
+        if ($comment->comment_approved == "1") {
+            return CommentStatus::APPROVE;
+        } elseif ($comment->comment_approved == "spam") {
+            return CommentStatus::SPAM;
+        } elseif ($comment->comment_approved == "trash") {
+            return CommentStatus::TRASH;
+        };
+        return CommentStatus::HOLD;
     }
     public function getCommentParent(object $comment): string | int | null
     {