+++ layers/Schema/packages/comments/src/FieldInterfaceResolvers/CommentableFieldInterfaceResolver.php
@@ -29,9 +29,11 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
     {
         return [
             'areCommentsOpen',
-            'commentCount',
             'hasComments',
+            'commentCount',
             'comments',
+            'unrestrictedCommentCount',
+            'unrestrictedComments',
         ];
     }
 
@@ -39,9 +41,11 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
     {
         $types = [
             'areCommentsOpen' => SchemaDefinition::TYPE_BOOL,
-            'commentCount' => SchemaDefinition::TYPE_INT,
             'hasComments' => SchemaDefinition::TYPE_BOOL,
+            'commentCount' => SchemaDefinition::TYPE_INT,
             'comments' => SchemaDefinition::TYPE_ID,
+            'unrestrictedCommentCount' => SchemaDefinition::TYPE_INT,
+            'unrestrictedComments' => SchemaDefinition::TYPE_ID,
         ];
         return $types[$fieldName] ?? parent::getSchemaFieldType($fieldName);
     }
@@ -50,10 +54,12 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
     {
         return match ($fieldName) {
             'areCommentsOpen',
+            'hasComments',
             'commentCount',
-            'hasComments'
+            'unrestrictedCommentCount'
                 => SchemaTypeModifiers::NON_NULLABLE,
-            'comments'
+            'comments',
+            'unrestrictedComments'
                 => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
             default
                 => parent::getSchemaFieldTypeModifiers($fieldName),
@@ -64,9 +70,11 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
     {
         $descriptions = [
             'areCommentsOpen' => $this->translationAPI->__('Are comments open to be added to the custom post', 'pop-comments'),
-            'commentCount' => $this->translationAPI->__('Number of comments added to the custom post', 'pop-comments'),
             'hasComments' => $this->translationAPI->__('Does the custom post have comments?', 'pop-comments'),
+            'commentCount' => $this->translationAPI->__('Number of comments added to the custom post', 'pop-comments'),
             'comments' => $this->translationAPI->__('Comments added to the custom post', 'pop-comments'),
+            'unrestrictedCommentCount' => $this->translationAPI->__('[Unrestricted] Number of comments added to the custom post', 'pop-comments'),
+            'unrestrictedComments' => $this->translationAPI->__('[Unrestricted] Comments added to the custom post', 'pop-comments'),
         ];
         return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($fieldName);
     }
@@ -81,6 +89,14 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
             'commentCount' => [
                 CommentFilterInputContainerModuleProcessor::class,
                 CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_CUSTOMPOST_COMMENTCOUNT
+            ],
+            'unrestrictedComments' => [
+                CommentFilterInputContainerModuleProcessor::class,
+                CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_CUSTOMPOST_ADMINCOMMENTS
+            ],
+            'unrestrictedCommentCount' => [
+                CommentFilterInputContainerModuleProcessor::class,
+                CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_CUSTOMPOST_ADMINCOMMENTCOUNT
             ],
             default => parent::getFieldDataFilteringModule($fieldName),
         };
@@ -98,6 +114,7 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
         ];
         switch ($fieldName) {
             case 'comments':
+            case 'unrestrictedComments':
                 $limitFilterInputName = FilterInputHelper::getFilterInputName([
                     CommonFilterInputModuleProcessor::class,
                     CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_LIMIT
@@ -117,6 +134,7 @@ class CommentableFieldInterfaceResolver extends AbstractQueryableSchemaFieldInte
                     ]
                 );
             case 'commentCount':
+            case 'unrestrictedCommentCount':
                 return $filterInputNameDefaultValues;
         }
         return parent::getFieldDataFilteringDefaultValues($fieldName);