+++ layers/Schema/packages/comments/src/FieldResolvers/CommentFieldResolver.php
@@ -18,7 +18,6 @@ use PoP\Hooks\HooksAPIInterface;
 use PoP\LooseContracts\NameResolverInterface;
 use PoP\Translation\TranslationAPIInterface;
 use PoPSchema\Comments\ComponentConfiguration;
-use PoPSchema\Comments\Constants\Status;
 use PoPSchema\Comments\ModuleProcessors\CommentFilterInputContainerModuleProcessor;
 use PoPSchema\Comments\TypeAPIs\CommentTypeAPIInterface;
 use PoPSchema\Comments\TypeResolvers\CommentTypeResolver;
@@ -68,10 +67,13 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
             'customPostID',
             'approved',
             'type',
+            'status',
             'parent',
             'date',
             'responses',
             'responseCount',
+            'unrestrictedResponses',
+            'unrestrictedResponseCount',
         ];
     }
 
@@ -86,10 +88,13 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
             'customPostID' => SchemaDefinition::TYPE_ID,
             'approved' => SchemaDefinition::TYPE_BOOL,
             'type' => SchemaDefinition::TYPE_STRING,
+            'status' => SchemaDefinition::TYPE_STRING,
             'parent' => SchemaDefinition::TYPE_ID,
             'date' => SchemaDefinition::TYPE_DATE,
             'responses' => SchemaDefinition::TYPE_ID,
             'responseCount' => SchemaDefinition::TYPE_INT,
+            'unrestrictedResponses' => SchemaDefinition::TYPE_ID,
+            'unrestrictedResponseCount' => SchemaDefinition::TYPE_INT,
         ];
         return $types[$fieldName] ?? parent::getSchemaFieldType($typeResolver, $fieldName);
     }
@@ -102,10 +107,13 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
             'customPostID',
             'approved',
             'type',
+            'status',
             'date',
-            'responseCount'
+            'responseCount',
+            'unrestrictedResponseCount'
                 => SchemaTypeModifiers::NON_NULLABLE,
-            'responses'
+            'responses',
+            'unrestrictedResponses'
                 => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
             default
                 => parent::getSchemaFieldTypeModifiers($typeResolver, $fieldName),
@@ -123,10 +131,13 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
             'customPostID' => $this->translationAPI->__('ID of the custom post to which the comment was added', 'pop-comments'),
             'approved' => $this->translationAPI->__('Is the comment approved?', 'pop-comments'),
             'type' => $this->translationAPI->__('Type of comment', 'pop-comments'),
+            'status' => $this->translationAPI->__('Status of the comment', 'pop-comments'),
             'parent' => $this->translationAPI->__('Parent comment (if this comment is a response to another one)', 'pop-comments'),
             'date' => $this->translationAPI->__('Date when the comment was added', 'pop-comments'),
             'responses' => $this->translationAPI->__('Responses to the comment', 'pop-comments'),
             'responseCount' => $this->translationAPI->__('Number of responses to the comment', 'pop-comments'),
+            'unrestrictedResponses' => $this->translationAPI->__('[Unrestricted] Responses to the comment', 'pop-comments'),
+            'unrestrictedResponseCount' => $this->translationAPI->__('[Unrestricted] Number of responses to the comment', 'pop-comments'),
         ];
         return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
     }
@@ -136,6 +147,8 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
         return match ($fieldName) {
             'responses' => [CommentFilterInputContainerModuleProcessor::class, CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_RESPONSES],
             'responseCount' => [CommentFilterInputContainerModuleProcessor::class, CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_RESPONSECOUNT],
+            'unrestrictedResponses' => [CommentFilterInputContainerModuleProcessor::class, CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_ADMINRESPONSES],
+            'unrestrictedResponseCount' => [CommentFilterInputContainerModuleProcessor::class, CommentFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_ADMINRESPONSECOUNT],
             'date' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
             default => parent::getFieldDataFilteringModule($typeResolver, $fieldName),
         };
@@ -145,6 +158,7 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
     {
         switch ($fieldName) {
             case 'responses':
+            case 'unrestrictedResponses':
                 $limitFilterInputName = FilterInputHelper::getFilterInputName([
                     CommonFilterInputModuleProcessor::class,
                     CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_LIMIT
@@ -203,6 +217,9 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
             case 'type':
                 return $this->commentTypeAPI->getCommentType($comment);
 
+            case 'status':
+                return $this->commentTypeAPI->getCommentStatus($comment);
+
             case 'parent':
                 return $this->commentTypeAPI->getCommentParent($comment);
 
@@ -214,8 +231,8 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
                 );
 
             case 'responses':
+            case 'unrestrictedResponses':
                 $query = [
-                    'status' => Status::APPROVED,
                     'parent-id' => $typeResolver->getID($comment),
                 ];
                 $options = array_merge(
@@ -227,8 +244,8 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
                 return $this->commentTypeAPI->getComments($query, $options);
 
             case 'responseCount':
+            case 'unrestrictedResponseCount':
                 $query = [
-                    'status' => Status::APPROVED,
                     'parent-id' => $typeResolver->getID($comment),
                 ];
                 $options = $this->getFilterDataloadQueryArgsOptions($typeResolver, $fieldName, $fieldArgs);
@@ -246,6 +263,7 @@ class CommentFieldResolver extends AbstractQueryableFieldResolver
 
             case 'parent':
             case 'responses':
+            case 'unrestrictedResponses':
                 return CommentTypeResolver::class;
         }