+++ layers/Schema/packages/comments/src/FieldResolvers/CustomPostFieldResolver.php
@@ -14,7 +14,6 @@ use PoP\Engine\CMS\CMSServiceInterface;
 use PoP\Hooks\HooksAPIInterface;
 use PoP\LooseContracts\NameResolverInterface;
 use PoP\Translation\TranslationAPIInterface;
-use PoPSchema\Comments\Constants\Status;
 use PoPSchema\Comments\FieldInterfaceResolvers\CommentableFieldInterfaceResolver;
 use PoPSchema\Comments\TypeAPIs\CommentTypeAPIInterface;
 use PoPSchema\Comments\TypeResolvers\CommentTypeResolver;
@@ -62,8 +61,8 @@ class CustomPostFieldResolver extends AbstractQueryableFieldResolver
     {
         return [
             'areCommentsOpen',
-            'commentCount',
             'hasComments',
+            'commentCount',
             'comments',
         ];
     }
@@ -100,16 +99,16 @@ class CustomPostFieldResolver extends AbstractQueryableFieldResolver
                 return $typeResolver->resolveValue($post, 'commentCount', $variables, $expressions, $options) > 0;
 
             case 'commentCount':
+            case 'unrestrictedCommentCount':
                 $query = [
-                    'status' => Status::APPROVED,
                     'customPostID' => $typeResolver->getID($post),
                 ];
                 $options = $this->getFilterDataloadQueryArgsOptions($typeResolver, $fieldName, $fieldArgs);
                 return $this->commentTypeAPI->getCommentCount($query, $options);
 
             case 'comments':
+            case 'unrestrictedComments':
                 $query = [
-                    'status' => Status::APPROVED,
                     'customPostID' => $typeResolver->getID($post),
                 ];
                 $options = array_merge(
@@ -128,6 +127,7 @@ class CustomPostFieldResolver extends AbstractQueryableFieldResolver
     {
         switch ($fieldName) {
             case 'comments':
+            case 'unrestrictedComments':
                 return CommentTypeResolver::class;
         }